Conflicts:
	osl-graph/osl-graph-core/src/test/java/org/openspotlight/graph/SLGraphTest.java
	osl-graph/osl-graph-test-data/src/main/java/org/openspotlight/graph/query/AbstractGeneralQueryTest.java

+++ osl-common/src/main/java/org/openspotlight/common/util/Files.java
@@ -118,11 +118,14 @@ public class Files {
      */
     private static void deleteDir( final File dir ) throws Exception {
         checkNotNull("dir", dir); //$NON-NLS-1$
-        for (final File f : dir.listFiles()) {
-            if (f.isFile()) {
-                deleteFile(f);
-            } else {
-                deleteDir(f);
+        final File[] files = dir.listFiles();
+        if (files != null) {
+            for (final File f : files) {
+                if (f.isFile()) {
+                    deleteFile(f);
+                } else {
+                    deleteDir(f);
+                }
             }
         }
         dir.delete();
